adding in new functionality
we now have a Iterate function which turns all typically
iterable objects into object accepted by the range command

we also have a Range function to create a generator
similar to the xrange python function.

we have used the iterate function in the existing Map and CMap
to make the funcitons cleaner and more readable.

This functionality adds a struct type called Pair, which is
in line with the c++ Pair object